Split apart setup, deploy and create vm. Currently setup is still working.  need to verify others
